level writer/classes reorg
