+++ fast64_internal/oot/oot_object.py
@@ -1,7 +1,7 @@
 from bpy.types import Object
 from ..utility import ootGetSceneOrRoomHeader
 from .data import OoT_Data
-from .oot_level_classes import OOTRoom
+from .scene.exporter.classes import OOTRoom
 
 
 def addMissingObjectToProp(roomObj: Object, headerIndex: int, objectKey: str):
