+++ fast64_internal/oot/scene/exporter/to_c/__init__.py
@@ -1,4 +1,4 @@
-from .scene import getIncludes, getSceneC
+from .scene import getIncludes, getSceneC, ootExportSceneToC
 from .scene_table_c import modifySceneTable, getDrawConfig
 from .spec import editSpecFile
 from .scene_folder import modifySceneFiles, deleteSceneFiles
