+++ fast64_internal/oot/scene/exporter/to_c/scene_header.py
@@ -1,7 +1,7 @@
 from .....utility import CData, indent
 from .....f3d.f3d_gbi import ScrollMethod, TextureExportSettings
 from ....oot_model_classes import OOTGfxFormatter
-from ....oot_level_classes import OOTScene, OOTLight
+from ..classes import OOTScene, OOTLight
 from .scene_pathways import getPathData
 from .actor import getTransitionActorList, getSpawnActorList, getSpawnList
 from .scene_commands import getSceneCommandList
