+++ fast64_internal/oot/scene/operators.py
@@ -1,17 +1,16 @@
 import bpy, os
 from bpy.path import abspath
-from bpy.types import Operator, UILayout
+from bpy.types import Operator
 from bpy.props import EnumProperty, IntProperty, StringProperty
 from bpy.utils import register_class, unregister_class
 from bpy.ops import object
 from mathutils import Matrix, Vector
 from ...f3d.f3d_gbi import DLFormat
 from ...utility import PluginError, raisePluginError, ootGetSceneOrRoomHeader
-from ..oot_utility import ExportInfo, sceneNameFromID, getEnumName
-from ..oot_level_writer import ootExportSceneToC
+from ..oot_utility import ExportInfo, sceneNameFromID
 from ..oot_constants import ootEnumMusicSeq, ootEnumSceneID
 from ..oot_level_parser import parseScene
-from .exporter.to_c import clearBootupScene, modifySceneTable, editSpecFile, deleteSceneFiles
+from .exporter.to_c import clearBootupScene, modifySceneTable, editSpecFile, deleteSceneFiles, ootExportSceneToC
 
 
 def ootRemoveSceneC(exportInfo):
